Fix linting violations
