+++ src/Calcine/Post/BasePostLocator.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Calcine\Post;
+
+abstract class BasePostLocator implements PostLocatorInterface
+{
+    public function isValidPath(string $path): bool
+    {
+        return (($post = $this->postForPath()) !== null);
+    }
+}
