+++ src/Calcine/Post/PostLocator.php
@@ -5,20 +5,6 @@ namespace Calcine\Post;
 use Calcine\Path;
 use Calcine\Post;
 
-interface PostLocatorInterface {
-    public function pathForPost(Post $post): string;
-    public function postForPath(string $path): ?Post;
-    public function isValidPath(string $path): bool;
-}
-
-abstract class BasePostLocator implements PostLocatorInterface
-{
-    public function isValidPath(string $path): bool
-    {
-        return (($post = $this->postForPath()) !== null);
-    }
-}
-
 final class PostLocator extends BasePostLocator
 {
     public function __construct(private string $rootPath, private array $posts)
